VERY BASIC quality filtering
Bot will now only work on messages with at least 15 characters. This should reduce the frequency of very short outputs.

In future it will be nice to have somewhere where I can set this number with a command, but I can't be bothered messing with the client object for now.